str: the number of strings passed to str_cat() should be a size_t

+++ include/jeffpc/str.h
@@ -83,7 +83,7 @@ extern struct str *str_alloc_static(const char *s);
 extern size_t str_len(const struct str *str);
 extern int str_cmp(const struct str *a, const struct str *b);
 extern struct str *str_dup(const char *s);
-extern struct str *str_cat(int n, ...);
+extern struct str *str_cat(size_t n, ...);
 extern struct str *str_printf(const char *fmt, ...)
 	__attribute__((format (printf, 1, 2)));
 extern struct str *str_vprintf(const char *fmt, va_list args);
